added ShowCourses.tsx, moved code from Courses to Homepage and ShowCourses, deleted Courses.
