+++ client/src/pages/Homepage.tsx
@@ -1,10 +1,30 @@
-import React from "react";
-import Courses from "../component/Courses";
+import React, { useState, useEffect } from "react";
+import { Course } from "../models/course";
+import agent from "../actions/agent";
+import { PaginatedCourse } from "../models/paginatedCourse";
+import ShowCourses from "../component/ShowCourses";
+import { Row } from "antd";
 
 const Homepage = () => {
+  const [data, setData] = useState<PaginatedCourse>();
+  
+  useEffect(() => {
+    agent.Courses.list().then((response) => {
+      setData(response);
+    });
+  }, []);
+  
   return (
-    <div>
-      <Courses />
+    <div className="course">
+      <div className="course__header">
+        <h1>What to learn next?</h1>
+        <h2>New Courses Just For You...</h2>
+      </div>
+      <Row gutter={[24, 32]}>
+        {data && data.data.map((course: Course, index: number) => {
+          return <ShowCourses key={index} course={course}/>
+        })}
+      </Row>
     </div>
   );
 };
